requirements.txt
flask==2.3.3
pyjwt==2.8.0
requests==2.31.0
lxml==4.9.3
pyyaml==6.0.1
